feat: adding lots of new updates
