+++ Px to Em/AppDelegate.swift
@@ -39,4 +39,3 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         }
     }
 }
-
