+++ Px to Em/PxToEm.swift
@@ -27,16 +27,9 @@ struct PxToEm: View {
     @State private var show_modal: Bool = false
     @State var show_toast: Bool = false
     
-//    @State private var baseText = "16"
-//    @State private var pixelText = "16"
-//    @State private var scaleText = "1.000"
     @State private var baseTextEmpty = ""
     @State private var pixelTextEmpty = ""
     @State private var scaleTextEmpty = ""
-    
-//    lazy var pixelInt = Double(pixelText) ?? 16
-//    lazy var baseInt = Double(baseText) ?? 16
-//    lazy var scaleInt = Double(scaleText) ?? 1.000
     
     func pxToEms(baseInt: Double, pixelInt: Double, scaleInt: Double) -> Double {
         let emValue = (pixelInt / baseInt) * scaleInt
