+++ Px to Em/TopBar.swift
@@ -52,6 +52,6 @@ struct TopBar : View {
 
 struct TopBar_Previews: PreviewProvider {
     static var previews: some View {
-        /*@START_MENU_TOKEN@*/Text("Hello, World!")/*@END_MENU_TOKEN@*/
+        EmptyView()
     }
 }
